Ошибки исправлены.

+++ Lab_1/src/main/java/com/vsu/data/inaccuracy/Accuracy.java
@@ -3,5 +3,5 @@ package com.vsu.data.inaccuracy;
 import java.util.List;
 
 public interface Accuracy {
-    double calcAccuracy(List<Integer> arrReal, List<Integer> arrCompress);
+    double calcAccuracy(List<Double> arrReal, List<Double> arrCompress);
 }
